requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
bokeh==1.1.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
cycler==0.10.0
cytoolz==0.10.0
dask==2.3.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
funcsigs==1.0.2
future==0.17.1
google==2.0.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
h5py==2.9.0
httplib2==0.13.1
hyperopt==0.2.2
idna==2.8
imageio==2.5.0
importlib-metadata==1.1.0
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
lera==0.2.3
leveldb==0.20
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.3.0
mmdnn==0.2.5
more-itertools==8.0.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.2
notebook==6.0.2
numpy==1.16.4
olefile==0.46
packaging==19.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.0
py-spy==0.3.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
Pygments==2.4.2
pymongo==3.9.0
pyparsing==2.4.2
pyrsistent==0.15.6
pytest==5.3.1
python-dateutil==2.8.0
python-gflags==3.1.2
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
ray==0.7.6
redis==3.3.11
requests==2.22.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.1.0
Send2Trash==1.5.0
setproctitle==1.1.10
six==1.12.0
sklearn==0.0
soupsieve==1.9.3
tensorboardX==1.8
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
torch==1.2.0
torch-lr-finder==0.0.1
torchvision==0.4.0a0+6b959ee
tornado==6.0.3
tqdm==4.34.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
